Refactor filterIncludedTargets in TuistCache
